The basic test now seems to work and some test runs detect duplicates.
Details:
Refined the test so it compares the timestamps, writes diagnostic info
to the log and reads all the messages we wrote. It's extremely slow,
only reading messages in batches of around 13. I'll investigate why it's
so slow and how we might usefully speed up the test.

The test has already found duplicate timestamps, however they seem to
happen in clusters. Also at the moment, when the duplicates occur
they're in the first few records, leaving the remaining records
unchecked. It'd be useful to enhance the checking to compare all the
records and determine the distribution/clustering of duplicates.

I've found how to enable logging for the tests and change the
granularity for logging so the tests can write more logs and Kafka
fewer. When Kafka and Zookeeper run with INFO logging there's lots of
potentially interesting detail, some may help us uncover weaknesses in
how we're using Kafka and how it behaves.