The duplicate timestamps appear frequently with these changes.
Increasing the log.segment.bytes has improved the performance
significantly. Increasing the number of records to write/check then
became more practical (rather than waiting minutes for the test to run)
and duplicates happen most test runs.

There's still plenty to do to hone the settings and see if we can now
make the tests more potent while also tuning kafka so it doesn't
generate duplicates (if that's practical).

I've opened an issue to track the extent of the duplicates
https://github.com/commercetest/kafka-unit/issues/2